Create theme.ts
