+++ arches/app/media/js/utils/create-vue-application.js
@@ -7,67 +7,11 @@ import StyleClass from 'primevue/styleclass';
 import ToastService from 'primevue/toastservice';
 import Tooltip from 'primevue/tooltip';
 
-import { definePreset } from '@primevue/themes';
 import { createApp } from 'vue';
 import { createGettext } from "vue3-gettext";
-import Aura from '@primevue/themes/aura';
 
 import arches from 'arches';
-
-export const ArchesPreset = definePreset(Aura, {
-    primitive: {
-        sky: {
-            950: '#2d3c4b',  // matches arches sidebar
-        },
-    },
-    semantic: {
-        primary: {
-            50: '{sky.50}',
-            100: '{sky.100}',
-            200: '{sky.200}',
-            300: '{sky.300}',
-            400: '{sky.400}',
-            500: '{sky.500}',
-            600: '{sky.600}',
-            700: '{sky.700}',
-            800: '{sky.800}',
-            900: '{sky.900}',
-            950: '{sky.950}',
-        },
-    },
-    components: {
-        button: {
-            root: {
-                label: {
-                    fontWeight: 600,
-                },
-            },
-        },
-        datatable: {
-            column: {
-                title: {
-                    fontWeight: 600,
-                },
-            },
-        },
-        splitter: {
-            handle: {
-                background: '{surface.500}',
-            },
-        },
-    },
-});
-
-const DEFAULT_THEME = {
-    theme: {
-        preset: ArchesPreset,
-        options: {
-            prefix: 'p',
-            darkModeSelector: 'system',
-            cssLayer: false
-        }
-    }
-};
+import { DEFAULT_THEME } from "@/arches/theme.ts";
 
 export default async function createVueApplication(vueComponent, themeConfiguration) {
     /**
